active-learning-nn/requirements.txt
torch>=2.2.0
scikit-learn>=1.3
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
scipy>=1.10
PyYAML>=6.0
tqdm>=4.66
tabulate>=0.9
# Dev
black>=24.0
pytest>=7.4
